LE-296: Move addIframeBackgroundImage and removeIframeBackgroundImage from Bulletin Service to Bulletin Util
